make pc 64bit

+++ autoprecompiles/src/blocks/mod.rs
@@ -45,7 +45,7 @@ impl<I> BasicBlock<I> {
 
 pub trait Program<I> {
     /// Returns the base program counter.
-    fn base_pc(&self) -> u32;
+    fn base_pc(&self) -> u64;
 
     /// Returns the step size of the program counter.
     fn pc_step(&self) -> u32;
